mathmentor/math_mentor/requirements.txt
anthropic>=0.40.0
langchain>=0.3.0
langchain-anthropic>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langgraph>=0.2.55
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0
streamlit>=1.40.0
streamlit-audiorecorder>=0.0.3
openai>=1.50.0
pillow>=10.4.0
python-dotenv>=1.0.0
pydantic>=2.8.0
numpy>=1.26.0
torch>=2.0.0
torchvision>=0.15.0
